Changed directory, Updated parser with the comment and new spec
For the import of parser in test file, i didn't succeed to import easily.
