+++ openfoodfacts_taxonomy_parser/tests/unit/test.txt
@@ -1,0 +1,37 @@
+# test taxonomy
+
+stopwords:fr: aux,au,de,le,du,la,a,et
+
+synonyms:en:passion fruit, passionfruit
+
+synonyms:fr:fruit de la passion, fruits de la passion, maracuja, passion
+
+en:yogurts, yoghurts
+fr:yaourts, yoghourts, yogourts
+
+<en:yogurts
+en:banana yogurts
+fr:yaourts à la banane
+
+<en:yogurts
+en:Passion fruit yogurts
+fr:yaourts au fruit de la passion
+
+<fr:yaourts au fruit de la passion
+fr:yaourts au fruit de la passion allégés
+
+# meat
+
+en:meat
+vegan:en:no
+carbon_footprint_fr_foodges_value:fr:10
+
+<en:meat
+en:fake-meat
+vegan:en:yes
+
+en:fake-stuff
+
+<en:fake-stuff
+<en:fake-meat
+en:fake-duck-meat
