+++ openfoodfacts_taxonomy_parser/tests/unit/test_parser_unit.py
@@ -1,0 +1,45 @@
+import pytest
+import re
+
+#importing parser
+import os,sys
+script_dir = os.path.dirname( '__file__' )
+mymodule_dir = os.path.join( script_dir, '..' , '..', 'parser')
+sys.path.append( mymodule_dir )
+import parser
+
+def test_normalized_filename():
+    x = parser.Parser()
+    normalizer = x.normalized_filename
+    name = normalizer("test")
+    assert name == "test.txt"
+    name = normalizer("test.txt")
+    assert name == "test.txt"
+    name = normalizer("t")
+    assert name == "t.txt"
+
+def test_fileiter():
+    x=parser.Parser()
+    file=x.file_iter("test.txt")
+    language_code_prefix = re.compile('[a-zA-Z][a-zA-Z]:')
+    for counter,(_,line) in enumerate(file):
+        assert line == '' or line[0]=="#" or ":" in line
+        if counter==26:
+            assert line == "carbon_footprint_fr_foodges_value:fr:10"
+    assert counter == 37
+
+def test_normalizing():
+    x=parser.Parser()
+    text="Numéro #1, n°1 des ¾ des Français*"
+    text=x.normalizing(text,"fr")
+    assert text == "numero-1-n-1-des-des-francais"
+    text="Randôm Languäge wìth àccénts"
+    normal_text=x.normalizing(text,"fr")
+    assert normal_text == "random-language-with-accents"
+    normal_text=x.normalizing(text,"de")
+    assert normal_text == "randôm-languäge-wìth-àccénts"
+
+
+if __name__ == "__main__":
+    pytest.main()
+
